Split services by portal sections. File with common ones
