+++ src/app/instantiate-dialog/instantiate-dialog.component.ts
@@ -3,7 +3,8 @@ import { MatDialogRef } from "@angular/material";
 import { MAT_DIALOG_DATA } from "@angular/material";
 import { NgForm } from "@angular/forms";
 
-import { ServiceManagementService } from "../shared/services/service-management/serviceManagement.service";
+import { ServiceManagementService } from "../shared/services/service-management/service-management.service";
+import { CommonService } from "../shared/services/common/common.service";
 
 @Component({
   selector: "app-instantiate-dialog",
@@ -19,16 +20,16 @@ export class InstantiateDialogComponent implements OnInit {
   locations = new Array();
 
   constructor(
-    private serviceManagementService: ServiceManagementService,
+    private commonService: CommonService,
     public dialogRef: MatDialogRef<InstantiateDialogComponent>,
     @Inject(MAT_DIALOG_DATA) public data: any,
-    private instantiateDialog: ServiceManagementService
+    private serviceManagementService: ServiceManagementService
   ) {}
 
   ngOnInit() {
     this.loading = true;
     setTimeout(() => {
-      this.serviceManagementService
+      this.commonService
         .requestVims()
         .then(response => {
           this.loading = false;
@@ -64,6 +65,10 @@ export class InstantiateDialogComponent implements OnInit {
   }
 
   instantiate(service) {
-    this.instantiateDialog.instantiate(service, this.ingress, this.egress);
+    this.serviceManagementService.instantiate(
+      service,
+      this.ingress,
+      this.egress
+    );
   }
 }
