+++ src/app/licences/licences.component.ts
@@ -3,7 +3,7 @@ import { MatTableDataSource } from "@angular/material";
 
 import { Router, ActivatedRoute } from "@angular/router";
 
-import { ServiceManagementService } from "../shared/services/service-management/serviceManagement.service";
+import { ServiceManagementService } from "../shared/services/service-management/service-management.service";
 import { DialogDataService } from "../shared/services/dialog/dialog.service";
 import { DataTransferService } from "../shared/services/service-management/dataTransfer.service";
